[!!!][TASK] Remove headline from Text and Image content types
Headlines should be used as simple content type and not mixed
with other types. Maybe this can be reintroduced later on but
currently the user interface does not allow clean removal of
the headline in Image or Text elements, therefor the property
is removed from those content types. It can easily be replaced
by Headline elements as seen in the demo page.

This is marked as breaking change as we currently do not offer
a node migration to update to this new content type definition.
A migration will probably follow at a later time.
If you need this change and still have content with headlines
you can revert the changes to TypoScript and adjust the
templates to output the headline but not make it editable.

Change-Id: Idd1f5235153da02d605f99975fcff6e62c70c302