[!!!][FEATURE] define node types in custom NodeTypes.yaml
This change moves the node type definition from the settings to an
extra NodeTypes.yaml file.

This change should only be merged together with the corresponding
other ones in the different packages.

In order to upgrade your node types, just move them from Settings.yaml
to the top-level of NodeTypes.yaml.